refactor: helper, config 패키지 분리
