+++ src/main/java/roomescape/controller/AuthController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.infrastructure.CookieExtractor;
+import roomescape.controller.helper.CookieExtractor;
 import roomescape.service.AuthService;
 import roomescape.service.dto.LoginCheckResponse;
 import roomescape.service.dto.LoginRequest;
